feat(nextbike/visualization): rename files + add first part of 2d
